Extracted notifications related logic into mixin file to reduce duplication. Moved utils.py from miteD/service to miteD/ as the methods there are also used in middleware. Added host/port parameter to the '@api' decorator so it's easier to customise these (e.g. to run 2 instances on same machine) Added notifications support in middleware layer
