+++ miteD/__init__.py
@@ -1,0 +1,1 @@
+from miteD.mixin.notifications import notification_handler
